[Maps] Add basic NP shim to Maps app and expose map config (#51683)
* Move plugin init calls to new plugin file, update refs and bind to legacy var. Assume ref for license init

* Bind server calls to __LEGACY var and pass to plugin setup

* Expose map config externally

* Pass in NP features and licensing plugins to replace some xpack_main functionality

* Pass through legacy server config for routes dependency

* Expose map config for legacy plugins

* Remove unused legacy refs

* Update license check to conform with latest NP license practices

* Clean up

* Revert to old license check. Handle NP license checking in separate PR

* Add back the license check file
