+++ x-pack/legacy/plugins/maps/index.js
@@ -4,21 +4,15 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { resolve } from 'path';
-import { initRoutes } from './server/routes';
-import { getEcommerceSavedObjects } from './server/sample_data/ecommerce_saved_objects';
-import { getFlightsSavedObjects } from './server/sample_data/flights_saved_objects.js';
-import { getWebLogsSavedObjects } from './server/sample_data/web_logs_saved_objects.js';
-import mappings from './mappings.json';
-import { checkLicense } from './check_license';
-import { migrations } from './migrations';
-import { watchStatusAndLicenseToInitialize } from
-  '../../server/lib/watch_status_and_license_to_initialize';
-import { initTelemetryCollection } from './server/maps_telemetry';
 import { i18n } from '@kbn/i18n';
-import { APP_ID, APP_ICON, createMapPath, MAP_SAVED_OBJECT_TYPE } from './common/constants';
+import { resolve } from 'path';
+import mappings from './mappings.json';
+import { migrations } from './migrations';
+import { initTelemetryCollection } from './server/maps_telemetry';
 import { getAppTitle } from './common/i18n_getters';
 import _ from 'lodash';
+import { MapPlugin } from './server/plugin';
+import { APP_ID, APP_ICON, createMapPath, MAP_SAVED_OBJECT_TYPE } from './common/constants';
 
 export function maps(kibana) {
 
@@ -108,111 +102,35 @@ export function maps(kibana) {
       }
       initTelemetryCollection(usageCollection, server);
 
-      const xpackMainPlugin = server.plugins.xpack_main;
-      let routesInitialized = false;
+      const coreSetup = server.newPlatform.setup.core;
+      const newPlatformPlugins = server.newPlatform.setup.plugins;
+      const pluginsSetup = {
+        featuresPlugin: newPlatformPlugins.features,
+        licensing: newPlatformPlugins.licensing,
+      };
 
-      xpackMainPlugin.registerFeature({
-        id: APP_ID,
-        name: i18n.translate('xpack.maps.featureRegistry.mapsFeatureName', {
-          defaultMessage: 'Maps',
-        }),
-        icon: APP_ICON,
-        navLinkId: APP_ID,
-        app: [APP_ID, 'kibana'],
-        catalogue: [APP_ID],
-        privileges: {
-          all: {
-            savedObject: {
-              all: [MAP_SAVED_OBJECT_TYPE, 'query'],
-              read: ['index-pattern']
-            },
-            ui: ['save', 'show', 'saveQuery'],
-          },
-          read: {
-            savedObject: {
-              all: [],
-              read: [MAP_SAVED_OBJECT_TYPE, 'index-pattern', 'query']
-            },
-            ui: ['show'],
-          },
-        }
-      });
-
-      watchStatusAndLicenseToInitialize(xpackMainPlugin, this,
-        async license => {
-          if (license && license.maps && !routesInitialized) {
-            routesInitialized = true;
-            initRoutes(server, license.uid);
-          }
-        });
-
-      xpackMainPlugin.info
-        .feature(this.id)
-        .registerLicenseCheckResultsGenerator(checkLicense);
-
-      const sampleDataLinkLabel = i18n.translate('xpack.maps.sampleDataLinkLabel', {
-        defaultMessage: 'Map'
-      });
-      server.addSavedObjectsToSampleDataset('ecommerce', getEcommerceSavedObjects());
-      server.addAppLinksToSampleDataset('ecommerce', [
-        {
-          path: createMapPath('2c9c1f60-1909-11e9-919b-ffe5949a18d2'),
-          label: sampleDataLinkLabel,
-          icon: APP_ICON
-        }
-      ]);
-      server.replacePanelInSampleDatasetDashboard({
-        sampleDataId: 'ecommerce',
-        dashboardId: '722b74f0-b882-11e8-a6d9-e546fe2bba5f',
-        oldEmbeddableId: '9c6f83f0-bb4d-11e8-9c84-77068524bcab',
-        embeddableId: '2c9c1f60-1909-11e9-919b-ffe5949a18d2',
-        embeddableType: 'map',
-        embeddableConfig: {
-          isLayerTOCOpen: false
+      // legacy dependencies
+      const __LEGACY = {
+        pluginRef: this,
+        config: server.config,
+        mapConfig() { return server.config().get('map'); },
+        route: server.route.bind(server),
+        plugins: {
+          elasticsearch: server.plugins.elasticsearch,
+          xpackMainPlugin: server.plugins.xpack_main
         },
-      });
-
-      server.addSavedObjectsToSampleDataset('flights', getFlightsSavedObjects());
-      server.addAppLinksToSampleDataset('flights', [
-        {
-          path: createMapPath('5dd88580-1906-11e9-919b-ffe5949a18d2'),
-          label: sampleDataLinkLabel,
-          icon: APP_ICON
-        }
-      ]);
-      server.replacePanelInSampleDatasetDashboard({
-        sampleDataId: 'flights',
-        dashboardId: '7adfa750-4c81-11e8-b3d7-01146121b73d',
-        oldEmbeddableId: '334084f0-52fd-11e8-a160-89cc2ad9e8e2',
-        embeddableId: '5dd88580-1906-11e9-919b-ffe5949a18d2',
-        embeddableType: MAP_SAVED_OBJECT_TYPE,
-        embeddableConfig: {
-          isLayerTOCOpen: true
+        savedObjects: {
+          getSavedObjectsRepository: server.savedObjects.getSavedObjectsRepository
         },
-      });
+        addSavedObjectsToSampleDataset: server.addSavedObjectsToSampleDataset,
+        addAppLinksToSampleDataset: server.addAppLinksToSampleDataset,
+        replacePanelInSampleDatasetDashboard: server.replacePanelInSampleDatasetDashboard,
+        injectUiAppVars: server.injectUiAppVars,
+        getInjectedUiAppVars: server.getInjectedUiAppVars
+      };
 
-      server.addSavedObjectsToSampleDataset('logs', getWebLogsSavedObjects());
-      server.addAppLinksToSampleDataset('logs', [
-        {
-          path: createMapPath('de71f4f0-1902-11e9-919b-ffe5949a18d2'),
-          label: sampleDataLinkLabel,
-          icon: APP_ICON
-        }
-      ]);
-      server.replacePanelInSampleDatasetDashboard({
-        sampleDataId: 'logs',
-        dashboardId: 'edf84fe0-e1a0-11e7-b6d5-4dc382ef7f5b',
-        oldEmbeddableId: '06cf9c40-9ee8-11e7-8711-e7a007dcef99',
-        embeddableId: 'de71f4f0-1902-11e9-919b-ffe5949a18d2',
-        embeddableType: MAP_SAVED_OBJECT_TYPE,
-        embeddableConfig: {
-          isLayerTOCOpen: false
-        },
-      });
-
-      server.injectUiAppVars(APP_ID, async () => {
-        return await server.getInjectedUiAppVars('kibana');
-      });
+      const mapPluginSetup = new MapPlugin().setup(coreSetup, pluginsSetup, __LEGACY);
+      server.expose('getMapConfig', mapPluginSetup.getMapConfig);
     }
   });
 }
